pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# -------------------------------------
# Poetry: Metadata
# -------------------------------------

[tool.poetry]
name = "aap-eda"
version = "0.1.0"
description = ""
authors = ["Red Hat, Inc. <info@ansible.com>"]
readme = "README.md"
packages = [{ include = "aap_eda", from = "src" }]

[tool.poetry.scripts]
aap-eda-manage = 'aap_eda.manage:main'

# -------------------------------------
# Poetry: Dependencies
# -------------------------------------
[tool.poetry.extras]
all = ["psycopg2"]
dev = ["psycopg2-binary"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
django = ">=4.2,<4.3"
djangorestframework = "*"
dynaconf = ">=3"
drf-spectacular = "*"
channels = { version = "*", extras = ["daphne"] }
psycopg2 = { version = "*", optional = true }
psycopg2-binary = { version = "*", optional = true }
django-rq = "*"
rq = "*"
django-filter = "*"
pydantic = "*"
cryptography = "*"
kubernetes = "*"
podman = ">=4.5"
rq-scheduler = "^0.10"
xxhash = "*"
psycopg = "*"
# Experimental LDAP Integration https://issues.redhat.com/browse/AAP-16938
# We are temporarily updating it to use latest from devel branch to keep consistency accross
# other AAP components
django-ansible-base = { git = "https://github.com/ansible/django-ansible-base.git" }
jinja2 = "*"

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-django = "*"
pytest-asyncio = "*"
requests = { version = "*", python = "<4.0" }
pytest-cov = "^4.1.0"
pytest-redis = "^3.0.2"
pytest-lazy-fixture = "^0.6.3"

[tool.poetry.group.lint.dependencies]
flake8 = "*"
isort = "*"
black = "*"
flake8-broken-line = { version = "*", python = "<4.0" }
flake8-string-format = "*"
# This is an experimental linter.
ruff = "*"
# The rull claims that the flake8 plugins listed below are re-implemented,
#   These plugins will remain included until it's verified.
pep8-naming = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-docstrings = "*"
flake8-eradicate = { version = "*", python = "<4.0" }
flake8-print = "*"

[tool.poetry.group.dev.dependencies]
ipython = "*"

# -------------------------------------
# Tools
# -------------------------------------

[tool.black]
line-length = 79
target-version = ["py39", "py310"]

[tool.isort]
profile = "black"
combine_as_imports = true
line_length = 79

[tool.ruff]
line-length = 79
select = [
    "E",
    "F",
    "D",   # flake8-docstrings
    "TID", # flake8-tidy-imports
]
extend-ignore = [
    "D1", #  Missing docstrings errors
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"src/aap_eda/core/migrations/*" = ["E501"]
"tests/**/*.py" = [
    "S101",   # Asserts allowed in tests
    "ARG",    # Fixtures are not always used explicitly
    "SLF001", # Call private methods in tests
]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.ruff.pydocstyle]
convention = "pep257"
